feat: remove deprecated metadata keys
